Test adding label with marker layer
